requirements.txt
aiohttp-jinja2==1.5
aiogram==2.20
environs==9.5.0
sqlalchemy
psycopg2
asyncpg
redis
implicit
numpy
scipy
pandas
html5lib
fake-useragent